Update NEUTRONRCF407 (#370)
* Update NEUTRONRCF407

USE_MAG is a cloud build option. When user would try another compass current configuration would need to include custom defines for other compass drivers.

* Update config.h

The board seems to have physical compass onboard so change it to allow other compass drivers when user decides to use another device. Likely the driver needs to be defined and or the ic2 address.