Update config.h
Per the manual, VTX power is controlled by PINIO on USER1. The config does not have PINIO configured for USER1, so it is impossible to power the VTX. This PR sets PINIO2 to USER1 so it is available in the modes tab. Note that PINIO1 is used to inactivate the on-board WiFi when armed. 

https://www.diatone.us/products/mb-mk4-h743-v2-fc