Update MLTYPHF4/config.h with spi vtx RTC6705 support
The Motolabs Typhoon F4 stack's VTX connects via RTC6705. The pins for this were defined in it's config, but named differently for unknown reasons. 

This adds defines for the correct variable names as well as declaring it's necessary include. 

Works with both full and cloud builds.

As seen in: https://discord.com/channels/868013470023548938/1269705725685399666